Refactor naming
