+++ src/main/java/com/tradeshift/models/HelloResult.java
@@ -1,4 +1,4 @@
-package com.tradeshift.model;
+package com.tradeshift.models;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
