+++ src/main/java/com/tradeshift/repositoreis/MessagesDAO.java
@@ -1,6 +1,6 @@
-package com.tradeshfit.model.dao;
+package com.tradeshift.repositoreis;
 
-import com.tradeshift.model.dto.MessageDTO;
+import com.tradeshift.models.dtos.MessageDTO;
 
 import java.util.Date;
 import java.util.List;
