+++ src/main/java/com/tradeshift/resources/HomeController.java
@@ -1,0 +1,10 @@
+package com.tradeshift.resources;
+
+import javax.ws.rs.Path;
+
+@Path("/")
+public class HomeController {
+
+    public void Index(){
+    }
+}
