+++ src/main/java/com/tradeshift/services/HelloService.java
@@ -1,4 +1,4 @@
-package com.tradeshift.service;
+package com.tradeshift.services;
 
 public interface HelloService {
     String formatName(String name) throws IllegalArgumentException;
