+++ src/test/java/com/tradeshfit/service/HelloEnglishTest.java
@@ -1,7 +1,7 @@
 package com.tradeshfit.service;
 
-import com.tradeshift.service.EnglishHelloService;
-import com.tradeshift.service.HelloService;
+import com.tradeshift.services.EnglishHelloService;
+import com.tradeshift.services.HelloService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.Assert;
